Remove generic classes
Added these in an attempt to match markup output by the Genesis featured page widget but it produces unpredictable results in themes inside and even moreso outside of the Genesis ecosystem.

We should revisit this when we have a way to filter output, where we can inject the markup we need when we need it.