Add new features to Post Grid block (#66)
* Add post class and ID to posts

* Add offset setting to post grid

* Add exclude sticky post option

* Add Post Grid Content panel

* Remove accent color from post grid links.

Fixes #53

* Add post excerpt length

* Refactor block structure

* Move inspector code to Inspector
* Clean up editor attributes

* Improve alt text on post grid

* Adds alt text to link if missing from image
* Adds screen reader text to read more link

Fixes #54

Thanks @BobbyBrooksATX for the a11y tips!

* Add content type selector

* Add markup to match Genesis widget

* Add post grid section title

* Add inline comments

* Fix wording

* Update readme files

* Remove FontAwesome from front end

Fixes #68

* Improve mobile styles

* Remove unnecessary const

* Remove unnecessary CPTs

Fixes #59

* Add FontAwesome back to editor styles for blocks

* Remove custom colors in favor of theme

* Add FontAwesome back to front end

Turns out this is needed on the front and back end for social icons in a few blocks. Will need to circle back to this and look at a better solution. Maybe just extracting the social icons instead of loading the whole FontAwesome.

* Add rel to links to account for core

* Remove custom button colors

* Remove custom color palettes in favor of theme

Fixes #52

* Revert fixes not related to postGrid

Reverts commits:

b1783a1e
ad160834
e93b6767
24670b0e
7645405d
29bfd51c
d90eb905
becf1412
9ea74889
320a7f5e

* Rebuild after reverting commits

* Remove generic classes

Added these in an attempt to match markup output by the Genesis featured page widget but it produces unpredictable results in themes inside and even moreso outside of the Genesis ecosystem.

We should revisit this when we have a way to filter output, where we can inject the markup we need when we need it.

* Add ab classes to header and more link

* Add sticky variable

* Improve display without featured image

* Add Post and Page Grid Markup section

This section allows the user to change the container tag, section title tag, and and post/page title tag to match their content hierarchy.

* Put section title behind toggle

* Remove H1 as an option

* Remove unused const and styles

* Improve excerpt

* Fix output of title

* Don’t show the title if toggled off

* Fix title toggle in editor

* Modify alt text behavior

* Remvoe alt from link

* Add aria and tabindex by default.

* Inline attributes for easier maintenance

* Remvoe unnecessary const and imports

* Tweak byline to be theme friendly

* Convert to WP_Query and remove sticky

* Regen POT

* Fix formatting

* phpcs fixes

* Change class to className

* Fix fontAwesome conflict

* Fix readme conflicts