Rebrand Getting Started page (#339)
* Rebrand GS page
- Split the page content into two files (AB/GPB)
- Apply conditional styling as needed

* Remove unused install URLS

* Remove FontAwesome dependency
- Replace with inline SVGs

* Update sidebar box heading color

* Conditionally set block category name.

Conditionally setting menu page name/title.

* Use Genesis font for menu icon

Needed to load the icon font to account for the different hover states and admin color schemes people use. We could conditionally load the icon font if Genesis is not detected (where it is already being loaded). Otherwise, this is temporary and the file size is very small.

* Remove unnecessary 'AB' from block titles

* Only display G icon when appropriate

* Remove “beta” language

* Replace doc links with new URL

* Fix alignment.

* Add atomic_blocks_settings_page_top hook and mark it as private.

* Fix modal helper bar

* Update feedback link

Co-authored-by: John Parris <john.parris@wpengine.com>