Fixes issue where automatic excerpts are not displayed.
The previous logic checked to see if the manual excerpt was empty before
generating one automatically. The problem was when checking for a
manual excerpt, it passed the data through the `the_excerpt` filter.
In cases where themes or plugins filtered `the_excerpt`, this caused
our empty check to assume there was a manual excerpt, therefore no
automatic excerpt was generated. An example of where this can
happen is with social sharing plugins that use the filter to add
sharing links/buttons.

This changes the logic so that it passes the data through
`the_excerpt` later, which avoids this issue.

Additional changes:
* Only checks for excerpts when the block is configured to show them.